Update FileManagerAlpha.java
I added in a more functional example we could use or build off of. This one is way more complicated though it seems to work for the most part with a gui. We don't need one so feel free to delete the code we don't need and we can salvage the rest.